Move EventAttendanceRepository to API layer
