+++ tmail-backend/jmap/extensions-api/src/main/scala/com/linagora/tmail/james/jmap/model/CalendarEventParse.scala
@@ -66,6 +66,10 @@ object CalendarEventParse {
 
 case class CalendarEventParsedList(value: List[CalendarEventParsed])
 
+case class CalendarEventNotDone(value: Map[UnparsedBlobId, SetError]) {
+  def merge(other: CalendarEventNotDone): CalendarEventNotDone = CalendarEventNotDone(this.value ++ other.value)
+}
+
 case class CalendarEventNotFound(value: Set[UnparsedBlobId]) {
   def merge(other: CalendarEventNotFound): CalendarEventNotFound = CalendarEventNotFound(this.value ++ other.value)
 }
